requirements.txt
kafka-python-ng==2.2.3
pytz
pyyaml
prometheus-client>=0.17.1
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
